+++ changelog.md
@@ -5,6 +5,7 @@ All notable outward-facing changes will be documented in this file.
 ## 1.1.0
 
 - Switch to UMD for broader compatibility
+- Remove circular dependency with `compress-tag`
 
 ## 1.0.0